Add remapping utilities
When in an "obfuscated" environment, the names aren't MCP ones, but
rather SRG. For this reason, there needs to be a way to convert one to
the other. In this case, we added the possibility to map from SRG to MCP
automatically in case it is needed. It's literally taken from ASMAPI,
but the goal is to make this cross-platform, as in the same function can
be used regardless of the Forge version (heck, technically Fermion's
goal could extend down to 1.7.10 in the coremods realm).

Also, added a test to let people see how hooking works, test whether it
does, and do the same with remapping.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>