Let Fermion launch plugins be shown in the Mod list
They don't get loaded as mods, the effect is simply, shall we say... for
aesthetic purpouses. The entry is added automatically and is treated as
a mod without ModContainer (basically a mod that cannot be constructed,
cannot receive events etc).

Also, this uses some interesting techniques of patching to reach across
class loaders. It may be interesting if you need to implement a pretty
down-to-Earth transformer (but I hope you won't: this is already so
prone to breaking that I may have to drop it in future releases).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>