Fix Launch Plugin File representation to now use a dummy file
Turns out that getFilePath() is supposed to be non-null and we are
initializing it with null. Not doing that in the mod constructor exposes
us to potential race conditions, so we are simply overriding the getter
as of now.

This fixes half of the issues Fermion had in dev, so I assume this is
also not prone to breaking with other mods.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>